requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
Brlapi==0.6.6
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==8.0.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
httplib2==0.9.2
idna==2.6
ImageHash==4.2.0
imageio==2.9.0
importlib==1.0.4
importlib-metadata==4.4.0
importlib-resources==3.0.0
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.0.0
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.9.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.16.0
sklearn==0.0
system-service==0.3
systemd-python==234
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.4.0+cpu
torchvision==0.7.0+cpu
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.0.35
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
